requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.2
django-mathfilters==1.0.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-recaptcha2==1.4.1
django-simple-captcha==0.5.18
django-summernote==0.8.20.0
django-taggit==4.0.0
idna==3.4
Pillow==10.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
